refactor: backend revamp
