+++ backend/config/db.js
@@ -1,5 +1,5 @@
-const mongoose = require("mongoose");
 require("dotenv").config();
+const mongoose = require("mongoose");
 
 const connectDB = async () => {
   try {
