+++ backend/controllers/authController.js
@@ -1,0 +1,57 @@
+const User = require("../models/User");
+const bcrypt = require("bcrypt");
+
+exports.register = async (req, res) => {
+  const { name, email, phoneNumber, password } = req.body;
+
+  try {
+    const existingUser = await User.findOne({
+      $or: [{ email }, { phoneNumber }],
+    });
+
+    if (existingUser) {
+      return res.status(400).json({ message: "User already exists" });
+    }
+
+    const hashedPassword = bcrypt.hashSync(password, 10);
+
+    const newUser = new User({
+      name,
+      email,
+      phoneNumber,
+      password: hashedPassword,
+      pin: "",
+    });
+
+    await newUser.save();
+
+    res.status(201).json({ message: "User registered successfully" });
+  } catch (err) {
+    console.error("Error during registration:", err);
+    res.status(500).json({ message: "Server error" });
+  }
+};
+
+exports.login = async (req, res) => {
+  try {
+    const { phoneNumber, password } = req.body;
+
+    const user = await User.findOne({ phoneNumber });
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    const isMatch = bcrypt.compareSync(password, user.password);
+    if (!isMatch) {
+      return res.status(401).json({ message: "Invalid number or password" });
+    }
+
+    res.status(200).json({
+      message: "Login successful",
+      user: { id: user._id, email: user.email },
+    });
+  } catch (error) {
+    console.error("Error logging in user:", error);
+    res.status(500).json({ message: "Server error" });
+  }
+};
