+++ backend/routes/authRoute.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const { login, register } = require("../controllers/authController");
+
+router.post("/register", register);
+router.post("/login", login);
+
+module.exports = router;
